chore: cr fixes
