+++ README.md
@@ -35,7 +35,7 @@ const summaryAgent = agent({
 Then, you can define and run your workflow.
 
 ```ts
-const translateFlow = chainFlow({
+const translateFlow = sequence({
   input: [
     {
       agent: 'translationAgent',
@@ -82,7 +82,7 @@ The patterns are inspired by Anthropic's agent patterns. You can learn more abou
 Use the `sequenceAgent` to chain multiple steps where output of one step becomes input for the next.
 
 ```typescript
-const translateAndSummarizeFlow = chainFlow({
+const translateAndSummarizeFlow = sequence({
   name: 'translateAndSummarize',
   input: [
     {
@@ -109,7 +109,7 @@ execute(translateAndSummarizeFlow, {
 Use the `oneOfAgent` to dynamically route to different execution paths based on conditions.
 
 ```typescript
-const routingFlow = routingflow({
+const routingFlow = routing({
   name: 'routeBasedOnSentiment',
   input: [
     {
@@ -138,7 +138,7 @@ execute(routingFlow, {
 Use the `parallelAgent` to run multiple steps concurrently and aggregate results.
 
 ```typescript
-const parallelAnalysisFlow = parallelFlow({
+const parallelAnalysisFlow = parallel({
   name: 'analyzeFromMultipleAngles',
   input: [
     {
@@ -170,7 +170,7 @@ execute(parallelAnalysisFlow, {
 Use the `optimizeAgent` to iteratively improve results based on specific criteria.
 
 ```typescript
-const optimizeFlow = evaluateFlow({
+const optimizeFlow = evaluate({
   name: 'improveWriting',
   input: {
     agent: 'writingAgent',
@@ -192,7 +192,7 @@ execute(optimizeFlow, {
 Use the `bestOfAllAgent` to generate multiple alternatives and pick the best one.
 
 ```typescript
-const bestOfFlow = bestOfFlow({
+const bestOfFlow = bestOfAll({
   name: 'generateBestResponse',
   criteria: 'Pick the response that is most helpful and concise'
   input: [
@@ -219,7 +219,7 @@ execute(bestOfFlow, {
 Use the `forEachAgent` to process a collection of items.
 
 ```typescript
-const processGithubIssuesFlow = forEachFlow({
+const processGithubIssuesFlow = forEach({
   input: [
     {
       agent: 'githubAgent',
